+++ sdk/messaging/azservicebus/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- `ReceivedMessage` can be converted to a `Message` for easier re-sending, using `ReceivedMessage.Message()`. PR#21472
+
 ### Breaking Changes
 
 ### Bugs Fixed